backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
httpx
supabase
python-multipart
fpdf2
python-jose
passlib==1.7.4
bcrypt==4.0.1